# Server/functions/src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Friend": {
				"properties": {
					"username": {
						"type": "string"
					},
					"uid": {
						"type": "string"
					},
					"availability": {
						"type": "string",
						"enum": [
							"Online",
							"Away",
							"Offline"
						]
					},
					"statusMessage": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"uid",
					"availability",
					"statusMessage"
				],
				"type": "object",
				"additionalProperties": true
			},
			"FriendRequest": {
				"properties": {
					"username": {
						"type": "string"
					},
					"uid": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"uid"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AddUserRequestDTO": {
				"properties": {
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"username",
					"id"
				],
				"type": "object"
			},
			"IUser": {
				"properties": {
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"puuid": {
						"type": "string",
						"nullable": true
					},
					"riotId": {
						"type": "string",
						"nullable": true
					},
					"rank": {
						"type": "string",
						"nullable": true
					},
					"preferredRoles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"friendsList": {
						"items": {
							"$ref": "#/components/schemas/Friend"
						},
						"type": "array"
					},
					"incomingRequests": {
						"items": {
							"$ref": "#/components/schemas/FriendRequest"
						},
						"type": "array"
					},
					"outgoingRequests": {
						"items": {
							"$ref": "#/components/schemas/FriendRequest"
						},
						"type": "array"
					},
					"blockedUsers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"availability": {
						"type": "string",
						"enum": [
							"Online",
							"Away",
							"Offline"
						]
					},
					"statusMessage": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"email",
					"puuid",
					"riotId",
					"rank",
					"preferredRoles",
					"friendsList",
					"incomingRequests",
					"outgoingRequests",
					"blockedUsers",
					"availability",
					"statusMessage"
				],
				"type": "object",
				"additionalProperties": true
			},
			"updateUserRequestDTO": {
				"properties": {
					"riotId": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"riotId",
					"id"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "functions",
		"description": "Cloud Functions for Firebase",
		"contact": {}
	},
	"paths": {
		"/user/add": {
			"post": {
				"operationId": "AddUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"createdUser": {
											"properties": {
												"statusMessage": {
													"type": "string"
												},
												"availability": {
													"type": "string",
													"enum": [
														"Online",
														"Away",
														"Offline"
													]
												},
												"blockedUsers": {
													"items": {
														"type": "string"
													},
													"type": "array"
												},
												"outgoingRequests": {
													"items": {
														"$ref": "#/components/schemas/FriendRequest"
													},
													"type": "array"
												},
												"incomingRequests": {
													"items": {
														"$ref": "#/components/schemas/FriendRequest"
													},
													"type": "array"
												},
												"friendsList": {
													"items": {
														"$ref": "#/components/schemas/Friend"
													},
													"type": "array"
												},
												"preferredRoles": {
													"items": {
														"type": "string"
													},
													"type": "array"
												},
												"rank": {
													"type": "string"
												},
												"riotId": {
													"type": "string"
												},
												"puuid": {
													"type": "string"
												},
												"email": {
													"type": "string"
												},
												"username": {
													"type": "string"
												}
											},
											"required": [
												"statusMessage",
												"availability",
												"blockedUsers",
												"outgoingRequests",
												"incomingRequests",
												"friendsList",
												"preferredRoles",
												"rank",
												"riotId",
												"puuid",
												"email",
												"username"
											],
											"type": "object"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"createdUser",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddUserRequestDTO"
							}
						}
					}
				}
			}
		},
		"/user/update": {
			"post": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatedUser": {
											"$ref": "#/components/schemas/IUser"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"updatedUser",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/updateUserRequestDTO"
							}
						}
					}
				}
			}
		},
		"/user/sendFriendRequest": {
			"post": {
				"operationId": "SendFriendRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "uid",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "targetUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/respondFriendRequest": {
			"patch": {
				"operationId": "RespondFriendRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "uid",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "incomingUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "accepted",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/user/removeFriend": {
			"delete": {
				"operationId": "RemoveFriend",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "uid",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "targetUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/toggleBlock": {
			"patch": {
				"operationId": "ToggleBlock",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "uid",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "targetUid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}